feat: make player and team pages indexable
