+++ app/layout.tsx
@@ -55,7 +55,7 @@ export const metadata: Metadata = {
     template: '%s | NBA 2K API',
   },
   description: 'The only free NBA 2K API. Access player ratings, team rosters, attributes, and badges via REST API. Developer-friendly documentation and playground.',
-  keywords: ['NBA 2K API', 'NBA 2K ratings API', 'NBA 2K player stats', '2K API', 'basketball API', 'NBA 2K player ratings', 'NBA 2K25 API'],
+  keywords: ['NBA 2K API', 'NBA 2K27 API', 'NBA 2K ratings API', 'NBA 2K player stats', '2K API', 'basketball API', 'NBA 2K player ratings'],
   authors: [{ name: 'Wilson Overfield', url: 'https://github.com/wkoverfield' }],
   creator: 'Wilson Overfield',
   openGraph: {
